+++ Example/src/component/HomeList/ListItem/index.js
@@ -1,6 +1,5 @@
 import React, {
   Component,
-  PropTypes,
 } from 'react';
 import {
   StyleSheet,
@@ -8,6 +7,7 @@ import {
   View,
   TouchableHighlight,
 } from 'react-native';
+import PropTypes from 'prop-types';
 
 const styles = StyleSheet.create({
   item: {