+++ CHANGELOG.md
@@ -10,6 +10,8 @@ n3：💚 不影响之前版本使用方式的更新（不需要用户适配）
 
 ## 0.18.0
 
+- 💚 import `PropTypes` from `prop-types` instead of `react`
+
 ## ActionSheet
 
 - 💚 new props `overlayAnimationDuration`：duration of Overlay animation